+++ database/migrations/2021_10_01_161241_create_avg_transportation_cost_to_incomes_table.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateAvgTransportationCostToIncomesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('avg_transportation_cost_to_incomes', function (Blueprint $table) {
+            $table->id();
+            $table->string('description');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('avg_transportation_cost_to_incomes');
+    }
+}