+++ resources/js/survey/views/survey/SocialData.vue
@@ -22,19 +22,29 @@
           Pendapatan per bulan
         </template>
       </card-survey>
-      <card-survey-slider v-if="step == 5" action="Lanjut" :height="6" :min="0" :max="100" :interval="1" :sliderTooltipFormat="'{value}%'" @onChange="input.avg_transportation_cost = $event" @onBack="backStep(4)" @onNext="nextStep(6)">
-        <template v-slot:question>
-          Berapa alokasi rata-rata biaya transportasi anda per bulan?
-        </template>
-        <template v-slot:info>
-          <span v-if="income.selected[0].max == null">≤</span>
-          <span v-if="income.selected[0].min != null"> Rp. {{income.selected[0].min * input.avg_transportation_cost/100 | currency}}</span>
+      <!-- udh ga pake percentage slider lagi -->
+      <b-col v-if="step == 5" lg="4" md="6" sm="8">
+        <b-card class="shadow-sm mb-3">
+          <div>
+            <p>
+              Berapa persentase rata – rata yang dialokasikan untuk biaya transportasi (tanpa transportasi udara) anda terhadap pendapatan? ​
+            </p>
+            <b-form-select :options="options.avg_transportation_cost" v-model="input.avg_transportation_cost" button-variant="outline-warning" buttons class="btn-block">
+              <template slot="first">
+                <option :value="null" disabled>-- Pilih Opsi--</option>
+              </template>
+            </b-form-select>
+          </div>
+        </b-card>
+        <div>
+          <b-btn variant="outline-secondary" @click="backStep(5)">
+            <i class="ri-arrow-left-circle-line align-middle"></i> Kembali
+          </b-btn>
+          <b-btn v-if="input.avg_transportation_cost != null" variant="primary" class="font-weight-bold float-right px-4" @click="nextStep(6)">Lanjut <i class="ri-arrow-right-circle-line align-middle"></i></b-btn>
+        </div>
+      </b-col>
 
-          <span v-if="income.selected[0].min != null && income.selected[0].max != null"> - </span>
-          <span v-if="income.selected[0].min == null">></span>
 
-          <span v-if="income.selected[0].max != null"> Rp. {{income.selected[0].max * input.avg_transportation_cost/100 | currency}}</span> </template>
-      </card-survey-slider>
       <card-survey v-if="step == 6" action="Lanjut" :options="options.vehicle" @onNext="nextStep(7)" @onBack="backStep(5)" @onChange="input.vehicle = $event">
         <template v-slot:question>
           Apa kendaraan yang punya?
@@ -146,6 +156,7 @@
           age: [],
           job: [],
           income: [],
+          avg_transportation_cost: [],
           vehicle: [],
           domicile_city: [],
           domicile_province: [],
@@ -254,6 +265,7 @@
             this.options.age = this.mutateKey(response.data.age);
             this.options.job = this.mutateKey(response.data.job);
             this.options.income = this.mutateKey(response.data.income);
+            this.options.avg_transportation_cost = this.mutateKey(response.data.avg_transportation_cost_to_income);
             this.options.domicile_province = this.mutateKey(response.data.province);
 
             this.options.vehicle = this.mutateKey(